Bump version of @zenfs/core to ^0.16.0 (#141)
- Previous version of zenfs-core had an engine requirement of >=22 which
was higher than the engine requirement for this library (18.18.2) which
caused issues with installing the library using yarn. (Yarn seemed to
error on engine mismatch but trying it with npm just seemed to log a
warning)

# Problem

In release 1.8.0 the browserfs dependency was replaced with zenfs
(#134). This ^0.14.0 version has a engine requirement of 22, which has
caused issues when I've tried to install this package in a node 18
project. The engine requirement for this library is 18.18.2.

Solution
========
In the 0.15.1 release of `@zenfs/core` they've added a polyfill and
reduced the engine requirement to node 16.
https://github.com/zen-fs/core/releases/tag/v0.15.1

I've bumped the version all the way to ^0.16.0 in case there were other
fixes along the way. The rest of the changelogs can be seen:
https://github.com/zen-fs/core/releases

You can see in the `package-lock.json` that the engine requirement has
changed from 22 to 16.

## Change summary:

- Bump version of @zenfs/core to ^0.16.0